Highlight patterns and customized clustering monkey patches
Style here is odd and should be reworked if we push to merge this fork
upstream.  These started out as monkey patches in our monolith.  I
wanted to quickly run them in other repos, so I forked and pulled them
in here.

`HighlightPatterns` lets you provide various text matching patterns and it
will colorize tables or rows that match.  Pattern matching is a tad
overengineered and is customized by an ENV var full of json.  If we
continue pushing this fork, we'll probably want to whittle down the
options.  I do like that this gives you arbitrarily many types of things
to colorize at least.

`CustomClustering` hijack's rails-erd's cluster feature.  Instead of
using the ruby module as a name space, it does some introspection.
Files are grouped by packwerk pack, gem, or code owner.  For a repo that
lacks intentional design, this has been a better way to group related
classes.